refactor: package structure -> [ebook, group, highlight, member, purchase]
