+++ chaekit-spring/src/main/java/qwerty/chaekit/controller/member/admin/AdminController.java
@@ -1,12 +1,12 @@
-package qwerty.chaekit.controller.admin;
+package qwerty.chaekit.controller.member.admin;
 import lombok.RequiredArgsConstructor;
 import org.springdoc.core.annotations.ParameterObject;
 import org.springframework.data.domain.Pageable;
 import org.springframework.web.bind.annotation.*;
-import qwerty.chaekit.dto.PublisherInfoResponse;
+import qwerty.chaekit.dto.member.PublisherInfoResponse;
 import qwerty.chaekit.dto.ebook.EbookListResponse;
-import qwerty.chaekit.dto.upload.EbookDownloadResponse;
-import qwerty.chaekit.dto.upload.EbookUploadRequest;
+import qwerty.chaekit.dto.ebook.upload.EbookDownloadResponse;
+import qwerty.chaekit.dto.ebook.upload.EbookUploadRequest;
 import qwerty.chaekit.global.response.ApiSuccessResponse;
 import qwerty.chaekit.service.AdminService;
 import qwerty.chaekit.service.EbookFileService;
