+++ chaekit-spring/src/main/java/qwerty/chaekit/dto/member/PublisherMemberResponse.java
@@ -1,4 +1,4 @@
-package qwerty.chaekit.dto;
+package qwerty.chaekit.dto.member;
 
 import lombok.Builder;
