+++ chaekit-spring/src/main/java/qwerty/chaekit/dto/member/UserJoinResponse.java
@@ -1,10 +1,11 @@
-package qwerty.chaekit.dto;
+package qwerty.chaekit.dto.member;
 
 import lombok.Builder;
 
 @Builder
-public record UserMemberResponse(
+public record UserJoinResponse(
         Long id,
+        String accessToken,
         String nickname,
         String username,
         String role
