+++ chaekit-spring/src/main/java/qwerty/chaekit/global/security/filter/login/LoginFilter.java
@@ -11,9 +11,9 @@ import org.springframework.security.core.Authentication;
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-import qwerty.chaekit.dto.LoginResponse;
+import qwerty.chaekit.dto.member.LoginResponse;
 import qwerty.chaekit.global.security.model.CustomUserDetails;
-import qwerty.chaekit.dto.LoginRequest;
+import qwerty.chaekit.dto.member.LoginRequest;
 import qwerty.chaekit.global.jwt.JwtUtil;
 import qwerty.chaekit.global.util.SecurityRequestReader;
 import qwerty.chaekit.global.util.SecurityResponseSender;
