+++ chaekit-spring/src/main/java/qwerty/chaekit/service/EbookFileService.java
@@ -7,8 +7,8 @@ import org.springframework.web.multipart.MultipartFile;
 import qwerty.chaekit.domain.ebook.Ebook;
 import qwerty.chaekit.domain.ebook.EbookRepository;
 import qwerty.chaekit.domain.member.publisher.PublisherProfile;
-import qwerty.chaekit.dto.upload.EbookDownloadResponse;
-import qwerty.chaekit.dto.upload.EbookUploadRequest;
+import qwerty.chaekit.dto.ebook.upload.EbookDownloadResponse;
+import qwerty.chaekit.dto.ebook.upload.EbookUploadRequest;
 import qwerty.chaekit.global.exception.BadRequestException;
 import qwerty.chaekit.global.exception.NotFoundException;
 import qwerty.chaekit.global.properties.AwsProperties;
