+++ chaekit-spring/src/main/java/qwerty/chaekit/service/PublisherJoinService.java
@@ -7,8 +7,8 @@ import qwerty.chaekit.domain.member.Member;
 import qwerty.chaekit.domain.member.enums.Role;
 import qwerty.chaekit.domain.member.publisher.PublisherProfile;
 import qwerty.chaekit.domain.member.publisher.PublisherProfileRepository;
-import qwerty.chaekit.dto.PublisherJoinRequest;
-import qwerty.chaekit.dto.PublisherJoinResponse;
+import qwerty.chaekit.dto.member.PublisherJoinRequest;
+import qwerty.chaekit.dto.member.PublisherJoinResponse;
 import qwerty.chaekit.global.exception.BadRequestException;
 import qwerty.chaekit.global.jwt.JwtUtil;
